requirements.txt
affine==2.3.0
alabaster==0.7.12
appdirs==1.4.4
astor==0.8.1
attrs==20.3.0
autoflake==1.4
Babel==2.9.0
bandit==1.6.2
black==20.8b1
certifi==2020.11.8
cftime==1.3.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
coverage==5.3
darglint==1.5.8
dict-deep==4.1.2
distlib==0.3.1
doc8==0.8.1
docutils==0.16
eradicate==1.0
filelock==3.0.12
flake8==3.8.4
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.3.0
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
future==0.18.2
GDAL==2.2.0
geoserver-restconfig==2.0.4.8
gisdata==0.5.4
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==4.3.21
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
netCDF4==1.5.4
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pathspec==0.8.1
pbr==5.5.1
pep8-naming==0.9.1
pika==1.1.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
python-dateutil==2.8.1
python-dev-tools==2020.9.10
pytz==2020.4
pyupgrade==2.7.4
PyYAML==5.3.1
rasterio==1.1.8
regex==2020.11.13
requests==2.25.0
restructuredtext-lint==1.3.2
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
snuggs==1.4.7
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stevedore==3.2.2
testfixtures==6.15.0
tokenize-rt==4.0.0
toml==0.10.2
tox==3.20.1
tox-travis==0.12
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.1
wemake-python-styleguide==0.14.1
xarray==0.16.1
